+++ unimport/statement.py
@@ -1,5 +1,4 @@
-from types import ModuleType
-from typing import List, NamedTuple, Optional
+from typing import List, NamedTuple
 
 
 class Name(NamedTuple):
@@ -7,15 +6,12 @@ class Name(NamedTuple):
     name: str
 
 
-class Import(NamedTuple):
-    lineno: int
-    name: str
-    module: Optional[ModuleType]
+class Import(Name):
+    pass
 
 
 class ImportFrom(NamedTuple):
     lineno: int
     name: str
     star: bool
-    module: Optional[ModuleType]
-    modules: List[str]
+    suggestions: List[str]